Added some HTML pages and a service directory which will hold service classes which handle html requests.
I added a home.html page along with a styles.css file for styling the html pages. I used bootstrap with the html pages to make it look more modern. I also added a prototype file for the service class which will handle the HTML requests such as GET races and UPDATE database.